+++ src/app/pages/demo-components/filter-components/demo/demo.component.ts
@@ -2,21 +2,18 @@ import { CommonModule } from '@angular/common';
 import { AfterViewInit, ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { ProductDataModel, SampleProductsData } from 'src/app/shared/constant/products';
-
-// import { selectedFilterSliderModelInterface, SelectedFilterInterface } from 'medes-ui/lib/mds-filter/mds-filter.interface';
-// import { MdsFilterCheckboxComponent, MdsFilterPipe, MdsFilterResetComponent, MdsFilterSwatchComponent, MdsHightlightPrismModule, MdsNumberUtils, MdsSearchPipe } from 'medes-ui';
-
-import { MdsFilterRangeSliderComponent } from 'projects/medes-ui/src/lib/mds-filter/filter-range-slider/filter-range-slider.component';
-import { SelectedFilterInterface, SelectedFilterSliderInterface } from 'projects/medes-ui/src/lib/mds-filter/mds-filter.interface';
-import { MdsFilterRangePipe } from 'projects/medes-ui/src/lib/mds-pipes-module/mds-filter-range.pipe';
-import { MdsFilterCheckboxComponent, MdsFilterPipe, MdsFilterResetComponent, MdsFilterSwatchComponent, MdsHightlightPrismModule, MdsNumberUtils, MdsSearchPipe } from 'projects/medes-ui/src/public-api';
+import { MdsHightlightPrismModule, MdsFilterModule, MdsPipesModule, MdsNumberUtils } from 'medes-ui';
+import { SelectedFilterInterface, SelectedFilterSliderInterface } from 'medes-ui/lib/mds-filter/mds-filter.interface';
+// Medes Team Only
+// import { MdsFilterModule, MdsHightlightPrismModule, MdsPipesModule, MdsNumberUtils } from 'projects/medes-ui/src/public-api';
+// import { SelectedFilterInterface, SelectedFilterSliderInterface } from 'projects/medes-ui/src/lib/mds-filter/mds-filter.interface';
 
 @Component({
   selector: 'mds-demo',
   templateUrl: './demo.component.html',
   styleUrls: ['./demo.component.scss'],
   standalone: true,
-  imports: [CommonModule, FormsModule, MdsHightlightPrismModule, MdsFilterCheckboxComponent, MdsFilterSwatchComponent, MdsFilterResetComponent, MdsFilterPipe, MdsSearchPipe, MdsFilterRangeSliderComponent, MdsFilterRangePipe]
+  imports: [CommonModule, FormsModule, MdsHightlightPrismModule, MdsFilterModule, MdsPipesModule]
 })
 export class DemoComponent implements OnInit, AfterViewInit {
   sampledata: ProductDataModel[] = [];